id採番ルール変更

+++ database/migrations/2021_06_22_115519_create_mem_statuses_table.php
@@ -14,7 +14,7 @@ class CreateMemStatusesTable extends Migration
     public function up()
     {
         Schema::create('mem_statuses', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->Integer('id');
             $table->string('status_name');
             $table->timestamps();
         });
