+++ database/migrations/2021_06_22_132551_create_gym_statuses_table.php
@@ -14,7 +14,7 @@ class CreateGymStatusesTable extends Migration
     public function up()
     {
         Schema::create('gym_statuses', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->Integer('id');
             $table->string('status_name');
             $table->timestamps();
         });
