+++ database/migrations/2021_06_28_102903_create_gym_types_table.php
@@ -14,7 +14,7 @@ class CreateGymTypesTable extends Migration
     public function up()
     {
         Schema::create('gym_types', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->Integer('id');
             $table->string('gym_type');
             $table->timestamps();
         });
